Synchronize dvb-apps with v4l-dvb
From: Andrea Odetti <mariofutire@gmail.com>

This patch (posted for the 3rd times) keeps dvb-apps in line with recent (and not so recent) changes
in v4l-dvb:

1) for 2 years it has not been possible to open the dvr more than once for read only.
After this change http://linuxtv.org/hg/v4l-dvb/rev/64edfcc88eeb the dvr can only be opened once.
The patch fixes a comment to the function dvbdemux_open_dvr

2) this changeset http://linuxtv.org/hg/v4l-dvb/rev/65699e8bc6f7 added the option
DMX_OUT_TSDEMUX_TAP to send the TS to the demux.
The patch enhances dvbdemux.h to use the DMX_OUT_TSDEMUX_TAP

3) a recent changest http://linuxtv.org/hg/v4l-dvb/rev/8389fb4e774c implemented DMX_SET_BUFFER for
the dvr.
tzap used to set the dvr buffer size to 1MB (this ioctl call used to be ignored) while the default
size is actually double (~2MB). I think the aim of the code was to make the buffer bigger and not to
shrink it. That buffer in my opinion should stay as it is.