Fix scan handling for EN 300468 charsets
It always bothered me that scan tables produced in Brazil
were full of "reverse interrogation" pseudo-chars, displayed
when some code violates UTF-8 charset.

This is due to the lack of properly interpreting the service provider
and service name charsets. There's a FIXME suggesting to use glibc
standard iconv functions in order to handle them.

According with EN 300 468, the first char (or the first 3 chars) may
specify the code to be used if they are different than code table 00
(Latin, ISO-6937 variant), and if the first character is < 0x20.

Unfortunately, not all DVB operators hounor it, so an extra parameter
were added to allow overriding.

However, at least here, the DVB-C operator doesn't honour the
charset table selection.

Due to that, a new option were added to scan, allowing to override
the charset default. So, in the case of Brazil, using a:
	-C ISO-8859-1
parameter, will properly produce the right channel description.

The specs allow the usage of certain control character sequences.
This is honoured by the descriptorcpy() function, for 1-byte char
tables (iso-8859 and iso-6937 tables).

Since I don't have access to dual-byte charset signals, I wouldn't
be able to test the conversions for iso-10646 table. So, I opted to
not handle 2-byte control chars (just like the previous code), but
it should be easy to do it, if someone is interested on doing it.

The code will output data by default in UTF-8 format, as this is
the charset used by modern distros, but user can use the -D option
to change the output to any other charset supported by iconv.
The supported charsets are given by:
	$ iconv -l

One interesting usage would be to force the output to ASCII, like:
	-D ASCII

As iconv is part of glibc, it should be available on any glibc-based
distribution. The iconv library is also present on other xNIX flavors.

As a future suggestion, someone could use LANG and LC_MESSAGES environment
vars to set the default for the output format.

Signed-off-by: Mauro Carvalho Chehab <mchehab@redhat.com>